+++ spec/unit/resource/api500/c7000/power_device_spec.rb
@@ -6,16 +6,15 @@
 #
 # Unless required by applicable law or agreed to in writing, software distributed
 # under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
-# CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
-# language governing permissions and limitations under the License.
+# CONDITIONS OF ANY KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations under the License.
 
-require_relative '../../api300/synergy/power_device'
+require 'spec_helper'
 
-module OneviewSDK
-  module API500
-    module Synergy
-      class PowerDevice < OneviewSDK::API500::C7000::PowerDevice
-      end
-    end
+RSpec.describe OneviewSDK::API500::C7000::PowerDevice do
+  include_context 'shared context'
+
+  it 'inherits from OneviewSDK::API300::C7000::PowerDevice' do
+    expect(described_class).to be < OneviewSDK::API300::C7000::PowerDevice
   end
 end